:recycle: [refactor]: 코드 제출 페이지 폴더 구조 수정
코드 제출 페이지를 Pages폴더에서 구현하지 않고 components/widget/problemsubmit 폴더로 분리 후 , 여기서 구현합니다
