+++ src/components/widget/problemsubmit/index.ts
@@ -1,0 +1,1 @@
+export { default as ProblemSubmit } from './ProblemSubmit';
